ajuste de dependencia de entidades

+++ carteirainteligente-api/src/main/java/br/com/carteirainteligente/api/controller/CategoryController.java
@@ -57,7 +57,11 @@ public class CategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Category> deleteCategory(@PathVariable Long id) {
+    public ResponseEntity<?> deleteCategory(@PathVariable Long id, @RequestBody Category category, BindingResult result) {
+        categoryValidator.validateDelete(id, result);
+        if (result.hasErrors()) {
+            return ResponseEntity.badRequest().body(result.getAllErrors());
+        }
         Category deletedCategory = categoryService.deleteCategory(id);
         return deletedCategory == null ? ResponseEntity.notFound().build() : ResponseEntity.ok(deletedCategory);
     }
